refactor: Remove MSGraphDb class and related usages; update dependencies in GraphSelectGuidancePlugin and IServiceCollectionExtensions
