+++ DevProxy.Plugins/Guidance/GraphSelectGuidancePlugin.cs
@@ -2,12 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using DevProxy.Abstractions.Data;
 using DevProxy.Abstractions.Proxy;
 using DevProxy.Abstractions.Plugins;
 using DevProxy.Abstractions.Utils;
 using Microsoft.Extensions.Logging;
-using Titanium.Web.Proxy.EventArguments;
 using System.Globalization;
+using Titanium.Web.Proxy.EventArguments;
 
 namespace DevProxy.Plugins.Guidance;
