+++ DevProxy/Commands/MsGraphDbCommand.cs
@@ -2,20 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using DevProxy.Abstractions.Utils;
+using DevProxy.Abstractions.Data;
 using System.CommandLine;
 
 namespace DevProxy.Commands;
 
 sealed class MsGraphDbCommand : Command
 {
-    private readonly ILogger _logger;
     private readonly MSGraphDb _msGraphDb;
 
-    public MsGraphDbCommand(ILogger<MsGraphDbCommand> logger, MSGraphDb msGraphDb) :
+    public MsGraphDbCommand(MSGraphDb msGraphDb) :
         base("msgraphdb", "Generate a local SQLite database with Microsoft Graph API metadata")
     {
-        _logger = logger;
         _msGraphDb = msGraphDb;
         ConfigureCommand();
     }
