+++ DevProxy/Extensions/IServiceCollectionExtensions.cs
@@ -3,9 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using DevProxy;
+using DevProxy.Abstractions.Data;
 using DevProxy.Abstractions.LanguageModel;
 using DevProxy.Abstractions.Proxy;
-using DevProxy.Abstractions.Utils;
 using DevProxy.Commands;
 using DevProxy.Proxy;
